Make unit tests green again
