+++ tests/Unit/Dto/ManualTest.php
@@ -4,6 +4,7 @@ namespace App\Tests\Unit\Dto;
 
 use App\Dto\Manual;
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\Finder\SplFileInfo;
 
 class ManualTest extends TestCase
 {
@@ -24,5 +25,32 @@ class ManualTest extends TestCase
         self::assertSame('en-us', $returnedManual->getLanguage());
         self::assertSame('c/typo3/cms-core/main/en-us', $returnedManual->getSlug());
         self::assertSame('main', $returnedManual->getVersion());
+    }
+
+    /**
+     * @test
+     */
+    public function returnsFilesWithSectionsForManual()
+    {
+        $filesRoot = implode(DIRECTORY_SEPARATOR, [
+            __DIR__,
+            'Fixtures',
+            'ParseDocumentationHTMLServiceTest',
+            'ReturnsFilesWithSectionsForManual',
+            'manual',
+        ]);
+
+        $manual = new Manual($filesRoot, 'title', 'type', 'main', 'en_us', 'slug');
+        $files = $manual->getFilesWithSections();
+        self::assertCount(3, $files);
+        $expectedFiles = [
+            $filesRoot . '/index.html',
+            $filesRoot . '/another.html',
+            $filesRoot . '/additional/index.html',
+        ];
+        foreach ($files as $file) {
+            /* @var $file SplFileInfo */
+            self::assertTrue(in_array((string)$file, $expectedFiles), 'Unexpected file: ' . $file);
+        }
     }
 }
